+++ tests/test_registration.py
@@ -1,13 +1,11 @@
-import pytest
-from click.testing import CliRunner
+from importlib import resources
 
 import numpy as np
+from click.testing import CliRunner
 
 import mesoscopy
 import mesoscopy.register as reg
-import mesoscopy.register.transform as trf
 import mesoscopy.resources as res
-import importlib.resources as resources
 
 
 def test_landmarks_affine():
@@ -15,14 +13,14 @@ def test_landmarks_affine():
 
 
 def test_load_preprocessed_h5(preproc_h5):
-    session_id, deltaf, ts = reg.load_preprocessed(preproc_h5)
+    session_id, deltaf, ts = reg.load_deltaf(preproc_h5)
     assert session_id == "preproc"
     assert deltaf.shape == (300, 40, 40)
     assert ts.shape == (300,)
 
 
 def test_load_preprocessed_nwb(preproc_nwb):
-    session_id, deltaf, ts = reg.load_preprocessed(preproc_nwb, nwb=True)
+    session_id, deltaf, ts = reg.load_deltaf(preproc_nwb, nwb=True)
     assert session_id == "session_1234"
     assert deltaf.shape == (300, 40, 40)
     assert ts.shape == (300,)
@@ -37,7 +35,7 @@ def test_update_nwb(nwbfile, preproc_h5):
 
 
 def test_register_landmarks_cli_default_points_h5(preproc_h5, output_dir):
-    mock_recording_landmarks = str(resources.files(mesoscopy.resources).joinpath("ccf_template_landmarks_140x142.csv"))
+    mock_recording_landmarks = str(resources.files(res).joinpath("ccf_template_landmarks_140x142.csv"))
     runner = CliRunner()
     result = runner.invoke(
         mesoscopy.cli,